Incremento do banco de dados, inserção das views do Perfil de Usuários e implementação da função estática para retorno de receitas livres de gluten ou lactose.

+++ src/models/Ingredient.java
@@ -6,9 +6,9 @@ public class Ingredient {
 
     private Food ingredient;
 
-    private int quantity;
+    private float quantity;
 
-    public Ingredient(Recipe recipe, Food ingredient, int quantity) {
+    public Ingredient(Recipe recipe, Food ingredient, float quantity) {
         this.recipe = recipe;
         this.ingredient = ingredient;
         this.quantity = quantity;
@@ -30,7 +30,7 @@ public class Ingredient {
         this.ingredient = ingredient;
     }
 
-    public int getQuantity() {
+    public float getQuantity() {
         return quantity;
     }
